+++ Source/SwiftLintFramework/Rules/Style/IdentifierNameRule.swift
@@ -34,9 +34,9 @@ struct IdentifierNameRule: ASTRule, ConfigurationProviderRule {
         }
 
         return validateName(dictionary: dictionary, kind: kind).map { name, offset in
-            guard !configuration.excluded.contains(name), let firstCharacter = name.first else {
-                return []
-            }
+            guard let firstCharacter = name.first else { return [] }
+
+            guard !configuration.shouldExclude(name: name) else { return [] }
 
             let isFunction = SwiftDeclarationKind.functionKinds.contains(kind)
             let description = Self.description