+++ Tests/SwiftLintFrameworkTests/IdentifierNameRuleTests.swift
@@ -2,6 +2,22 @@
 import XCTest
 
 class IdentifierNameRuleTests: XCTestCase {
+    func testIdentifierNameWithExcluded() {
+        let baseDescription = IdentifierNameRule.description
+        let nonTriggeringExamples = baseDescription.nonTriggeringExamples + [
+            Example("let Apple = 0"),
+            Example("let some_apple = 0"),
+            Example("let Test123 = 0")
+        ]
+        let triggeringExamples = baseDescription.triggeringExamples + [
+            Example("let ap_ple = 0"),
+            Example("let AppleJuice = 0")
+        ]
+        let description = baseDescription.with(nonTriggeringExamples: nonTriggeringExamples,
+                                               triggeringExamples: triggeringExamples)
+        verifyRule(description, ruleConfiguration: ["excluded": ["Apple", "some.*", ".*\\d+.*"]])
+    }
+
     func testIdentifierNameWithAllowedSymbols() {
         let baseDescription = IdentifierNameRule.description
         let nonTriggeringExamples = baseDescription.nonTriggeringExamples + [