+++ Tests/SwiftLintFrameworkTests/TypeNameRuleTests.swift
@@ -2,6 +2,22 @@
 import XCTest
 
 class TypeNameRuleTests: XCTestCase {
+    func testTypeNameWithExcluded() {
+        let baseDescription = TypeNameRule.description
+        let nonTriggeringExamples = baseDescription.nonTriggeringExamples + [
+            Example("class apple {}"),
+            Example("struct some_apple {}"),
+            Example("protocol test123 {}")
+        ]
+        let triggeringExamples = baseDescription.triggeringExamples + [
+            Example("enum ap_ple {}"),
+            Example("typealias appleJuice = Void")
+        ]
+        let description = baseDescription.with(nonTriggeringExamples: nonTriggeringExamples,
+                                               triggeringExamples: triggeringExamples)
+        verifyRule(description, ruleConfiguration: ["excluded": ["apple", "some.*", ".*st\\d+.*"]])
+    }
+
     func testTypeNameWithAllowedSymbols() {
         let baseDescription = TypeNameRule.description
         let nonTriggeringExamples = baseDescription.nonTriggeringExamples + [